A commit with regrets
I started this proj with js, and in between the journey I thought of ts and developed some logics in it, but it was too late. The core compiling  part was in js so i needed to make that also to ts , I couldnt  afford the time for.
It felt so heartbrooking to rewrite ts to js ones. So just attaching the ts files also to ease my regret.
Now electron will takje and store the data coming from go in a doubly linked list. Not sure its effective but its an experimental to try out this data structure since its my first time.
Bye